render keyboard when search field clicked and keyboard disappears onblur
