+++ package.json
@@ -18,6 +18,7 @@
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.22.3",
     "react-scripts": "5.0.1",
+    "react-simple-keyboard": "^3.7.107",
     "typescript": "^4.9.5",
     "web-vitals": "^2.1.4"
   },
